requirements.txt
paste
falcon==0.3.0
gunicorn>=19.2.0
keystonemiddleware
oslo.config>=1.2.1
oslo.middleware
oslo.log
oslo.serialization
oslo.utils
pastedeploy>=1.3.3
pbr>=1.6.0,<2.0
six>=1.9.0
kafka-python>=0.9.3,<0.9.4
simplejson>=3.8.0
simport